This set of code changes includes updates to the genesis state definition, the addition of new functions and parameters related to non-performance and slashing in the servicer keeper module, and updates to the fee structure in the types package.
In the servicer keeper module, new functions and parameters have been added to handle non-performance by validators, including the `MaxNonPerformantBlocks`, `MinScore`, and `SlashFractionBadPerformance` parameters. These parameters are used to determine when a validator's stake should be slashed due to non-performance. The `handleValidatorSignature` function has also been updated to take these new parameters into account.

In the types package, the fee for the report card message has been defined and the minimum number of proofs required has been increased.

These changes are likely part of ongoing development and testing of the non-performance and slashing features in the servicer keeper module.